requirements-dev.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
